Move heap into its own internal package
