+++ pkg/scheduler/util/BUILD
@@ -10,7 +10,6 @@ go_test(
     name = "go_default_test",
     srcs = [
         "error_channel_test.go",
-        "heap_test.go",
         "utils_test.go",
     ],
     embed = [":go_default_library"],
@@ -28,17 +27,14 @@ go_library(
     srcs = [
         "clock.go",
         "error_channel.go",
-        "heap.go",
         "utils.go",
     ],
     importpath = "k8s.io/kubernetes/pkg/scheduler/util",
     deps = [
         "//pkg/api/v1/pod:go_default_library",
         "//pkg/scheduler/apis/extender/v1:go_default_library",
-        "//pkg/scheduler/metrics:go_default_library",
         "//staging/src/k8s.io/api/core/v1:go_default_library",
         "//staging/src/k8s.io/apimachinery/pkg/apis/meta/v1:go_default_library",
-        "//staging/src/k8s.io/client-go/tools/cache:go_default_library",
         "//vendor/k8s.io/klog:go_default_library",
     ],
 )
